converting more funcs to helpers
